% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{docker_command}
\alias{docker_command}
\title{Run a Docker command}
\usage{
docker_command(args)
}
\arguments{
\item{args}{A single argument or vector of arguments to pass to [system2()].}
}
\value{
The output of the given command as a string.
}
\description{
Run a Docker command
}
\examples{
\dontrun{
  docker_command(c("info", "--format '{{json .}}'")
}
}
