% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{docker_info}
\alias{docker_info}
\title{Retrieve system Docker information}
\usage{
docker_info(simplify = TRUE)
}
\arguments{
\item{simplify}{Logical; Whether to simplify the json structure (e.g.
coerce list of lists elements into a data.frame) or return as one big list.}
}
\value{
A list containing Docker system information
}
\description{
Retrieve system Docker information
}
