% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jetty-package.R
\docType{package}
\name{jetty-package}
\alias{jetty}
\alias{jetty-package}
\title{jetty: Execute R in a 'Docker' Context}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'jetty' is to execute R functions and code snippets in an isolated R subprocess within a 'Docker' container and return the evaluated results to the local R session. 'jetty' can install necessary packages at runtime and seamlessly propagates errors and outputs from the 'Docker' subprocess back to the main session. 'jetty' is primarily designed for sandboxed testing and quick execution of example code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmolitor/jetty}
  \item \url{http://www.dmolitor.com/jetty/}
  \item Report bugs at \url{https://github.com/dmolitor/jetty/issues}
}

}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com} [copyright holder]

}
\keyword{internal}
