% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{docker_command}
\alias{docker_command}
\title{Run a Docker command}
\usage{
docker_command(args)
}
\arguments{
\item{args}{A single argument or vector of arguments to pass
to \code{\link[=system2]{system2()}}}
}
\value{
The output of the given command as a string.
}
\description{
Execute a function or code block within the context of a Docker container
and return the results to the local R session.
}
\examples{
\dontrun{
docker_info <- docker_command(c("info", "--format '{{json .}}'")
}
}
