% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jetty-package.R
\docType{package}
\name{jetty-package}
\alias{jetty}
\alias{jetty-package}
\title{jetty: Execute R in a Docker Context}
\description{
Execute functions and expressions within the context of a Docker container and return results to the local R session.
}
\keyword{internal}
