% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jetty-package.R
\docType{package}
\name{jetty-package}
\alias{jetty}
\alias{jetty-package}
\title{jetty: Execute R in a Docker Context}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Execute functions and expressions within the context of a Docker container and return results to the local R session.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmolitor/jetty}
  \item Report bugs at \url{https://github.com/dmolitor/jetty/issues}
}

}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com} [copyright holder]

}
\keyword{internal}
