% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Execute an R expression inside a Docker container}
\usage{
run(func, image = paste0("r-base:", r_version()), debug = FALSE)
}
\arguments{
\item{func}{Function object or expression to be executed in the R session
within the Docker container. Package functions should be referenced using
the `::` notation and only packages installed in the Docker container are
accessible.}

\item{image}{A string in the `image:tag` format specifying either a local
Docker image or an image available on DockerHub. Default image is
`r-base:your.r.version` where your R version is determined from your local
R session.}

\item{debug}{A logical indicating whether to print out the commands that are
being executed via the shell. This is mostly helpful to see what is
happening when things start to error.}
}
\value{
Value of the evaluated expression.
}
\description{
This function is somewhat similar in spirit to
\code{\link[callr:r]{callr::r()}} in that the user can pass
a function (or an expression) to be evaluated. This expression will
be executed within the context of a Docker container and the result will be
returned within the local R session.
}
\details{
It is important to note that most side effects, e.g. printing and plotting,
will be lost since these are being screamed into the void of the Docker
container. However, the local file system `path.expand("~")` and
\code{\link[=tempdir]{tempdir()}} are mounted to the Docker container, so
any side effects that involve writing to the local file system or the
temp directory will work as expected.
}
\section{Error handling}{


\link{jetty} will propagate errors from the Docker process to the main process.
That is, if an error is thrown in the Docker process, an error with the same
message will be thrown in the main process. However, because of the
somewhat isolated nature of the local process and the Docker process,
calling functions such as \code{\link[=traceback]{traceback()}} and
\code{\link[rlang:last_trace]{rlang::last_trace()}} will, unfortunately,
not print the callstack of the uncaught error as that has (in its current
form) been lost in the Docker void.
}

\examples{
\dontrun{
run(
  {
    mtcars <- mtcars |>
      transform(cyl = as.factor(cyl))
    model <- lm(mpg ~ ., data = mtcars)
    model
  }
)
}

}
